+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Smartness.java
@@ -1,13 +1,17 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
+import insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.IEnchantmentTooltip;
 import insane96mcp.insanelib.util.MathHelper;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.LootBonusEnchantment;
 
-public class Smartness extends Enchantment {
+public class Smartness extends Enchantment implements IEnchantmentTooltip {
     public Smartness() {
         super(Rarity.RARE, EnchantmentsFeature.WEAPONS_CATEGORY, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
     }
@@ -28,7 +32,16 @@ public class Smartness extends Enchantment {
         return super.checkCompatibility(enchantment) && !(enchantment instanceof LootBonusEnchantment);
     }
 
+    public static float getBonusExperience(int lvl) {
+        return lvl * 0.5f;
+    }
+
     public static int getIncreasedExperience(RandomSource random, int lvl, int experience) {
-        return MathHelper.getAmountWithDecimalChance(random, experience * (1 + lvl * .5f));
+        return MathHelper.getAmountWithDecimalChance(random, experience * (1f + getBonusExperience(lvl)));
+    }
+
+    @Override
+    public Component getTooltip(ItemStack stack, int lvl) {
+        return Component.translatable(this.getDescriptionId() + ".tooltip", getBonusExperience(lvl) * 100f).withStyle(ChatFormatting.DARK_PURPLE);
     }
 }